+++ product-service/src/main/java/io/github/lucasrech/productservice/domain/manufacturer/ManufacturerRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface ManufacturerRepository {
     Manufacturer save(Manufacturer manufacturer);
-    Optional<Manufacturer> findById(Integer id);
+    Optional<Manufacturer> findById(Short id);
     Page<Manufacturer> findAll(Pageable pageable);
     Optional<Manufacturer> findByCnpj(String cnpj);
     boolean existsByCnpj(String cnpj);
